+++ stac_fastapi/sqlalchemy/stac_fastapi/sqlalchemy/models/links.py
@@ -1,4 +1,3 @@
-# encoding: utf-8
 """link helpers."""
 
 from typing import Dict, List
